refactor: PaymentService 클래스의 패키지 변경
store 에서 store.service 패키지로 변경
